bigint

+++ mountainsort/src/isosplit5/isocut5.cpp
@@ -5,28 +5,28 @@
 #include <stdio.h>
 
 namespace ns_isocut5 {
-void copy_samples(int N, float* out, float* in);
-double sum(int N, float* X);
-int find_min_index(int N, float* X);
-int find_max_index(int N, float* X);
-double compute_ks4(int N, float* counts1, float* counts2);
-double compute_ks5(int* critical_range_min, int* critical_range_max, int N, float* counts1, float* counts2, int peak_index);
-void debug_print_array(int N, float* X);
+void copy_samples(bigint N, float* out, float* in);
+double sum(bigint N, float* X);
+bigint find_min_index(bigint N, float* X);
+bigint find_max_index(bigint N, float* X);
+double compute_ks4(bigint N, float* counts1, float* counts2);
+double compute_ks5(bigint* critical_range_min, bigint* critical_range_max, bigint N, float* counts1, float* counts2, bigint peak_index);
+void debug_print_array(bigint N, float* X);
 }
 
-void isocut5_mex(double* dipscore, double* cutpoint, int N, double* samples)
+void isocut5_mex(double* dipscore, double* cutpoint, bigint N, double* samples)
 {
     *dipscore = 0;
     *cutpoint = 0;
     isocut5_opts opts;
     float* samplesf = (float*)malloc(sizeof(float) * N);
-    for (int i = 0; i < N; i++)
+    for (bigint i = 0; i < N; i++)
         samplesf[i] = samples[i];
     isocut5(dipscore, cutpoint, N, samplesf, opts);
     free(samplesf);
 }
 
-void isocut5(double* dipscore_out, double* cutpoint_out, int N, float* samples, isocut5_opts opts)
+void isocut5(double* dipscore_out, double* cutpoint_out, bigint N, float* samples, isocut5_opts opts)
 {
     float* samples_sorted = (float*)malloc(sizeof(float) * N);
 
@@ -37,33 +37,33 @@ void isocut5(double* dipscore_out, double* cutpoint_out, int N, float* samples, 
         jisotonic5_sort(N, samples_sorted, samples);
 
     float num_bins_factor = 1;
-    int num_bins = ceil(sqrt(N * 1.0 / 2) * num_bins_factor);
+    bigint num_bins = ceil(sqrt(N * 1.0 / 2) * num_bins_factor);
 
-    int num_bins_1 = ceil(num_bins / 2);
-    int num_bins_2 = num_bins - num_bins_1;
-    int num_intervals = num_bins_1 + num_bins_2;
+    bigint num_bins_1 = ceil(num_bins / 2);
+    bigint num_bins_2 = num_bins - num_bins_1;
+    bigint num_intervals = num_bins_1 + num_bins_2;
     float* intervals = (float*)malloc(sizeof(float) * num_intervals);
-    for (int i = 0; i < num_bins_1; i++)
+    for (bigint i = 0; i < num_bins_1; i++)
         intervals[i] = i + 1;
-    for (int i = 0; i < num_bins_2; i++)
+    for (bigint i = 0; i < num_bins_2; i++)
         intervals[num_intervals - 1 - i] = i + 1;
     float alpha = (N - 1) / ns_isocut5::sum(num_intervals, intervals);
-    for (int i = 0; i < num_intervals; i++)
+    for (bigint i = 0; i < num_intervals; i++)
         intervals[i] *= alpha;
-    int N_sub = num_intervals + 1;
+    bigint N_sub = num_intervals + 1;
     float* inds = (float*)malloc(sizeof(float) * N_sub);
     inds[0] = 0;
-    for (int i = 0; i < num_intervals; i++)
+    for (bigint i = 0; i < num_intervals; i++)
         inds[i + 1] = inds[i] + intervals[i];
     float* X_sub = (float*)malloc(sizeof(float) * N_sub);
-    for (int i = 0; i < N_sub; i++)
-        X_sub[i] = samples_sorted[(int)inds[i]];
+    for (bigint i = 0; i < N_sub; i++)
+        X_sub[i] = samples_sorted[(bigint)inds[i]];
     float* densities = (float*)malloc(sizeof(float) * (N_sub - 1));
     float* spacings = (float*)malloc(sizeof(float) * (N_sub - 1));
     float* multiplicities = (float*)malloc(sizeof(float) * (N_sub - 1));
-    for (int i = 0; i < N_sub - 1; i++) {
+    for (bigint i = 0; i < N_sub - 1; i++) {
         spacings[i] = X_sub[i + 1] - X_sub[i];
-        multiplicities[i] = ((int)inds[i + 1]) - ((int)inds[i]);
+        multiplicities[i] = ((bigint)inds[i + 1]) - ((bigint)inds[i]);
         densities[i] = multiplicities[i] / spacings[i];
     }
 
@@ -71,25 +71,25 @@ void isocut5(double* dipscore_out, double* cutpoint_out, int N, float* samples, 
     float* densities_resid = (float*)malloc(sizeof(float) * (N_sub - 1));
     float* densities_unimodal_fit_times_spacings = (float*)malloc(sizeof(float) * (N_sub - 1));
     jisotonic5_updown(N_sub - 1, densities_unimodal_fit, densities, multiplicities);
-    for (int i = 0; i < N_sub - 1; i++)
+    for (bigint i = 0; i < N_sub - 1; i++)
         densities_resid[i] = densities[i] - densities_unimodal_fit[i];
-    for (int i = 0; i < N_sub - 1; i++)
+    for (bigint i = 0; i < N_sub - 1; i++)
         densities_unimodal_fit_times_spacings[i] = densities_unimodal_fit[i] * spacings[i];
-    int critical_range_min, critical_range_max;
-    int peak_index = ns_isocut5::find_max_index(N_sub - 1, densities_unimodal_fit);
+    bigint critical_range_min, critical_range_max;
+    bigint peak_index = ns_isocut5::find_max_index(N_sub - 1, densities_unimodal_fit);
     *dipscore_out = ns_isocut5::compute_ks5(&critical_range_min, &critical_range_max, N_sub - 1, multiplicities, densities_unimodal_fit_times_spacings, peak_index);
-    int critical_range_length = critical_range_max - critical_range_min + 1;
+    bigint critical_range_length = critical_range_max - critical_range_min + 1;
 
     float* densities_resid_on_critical_range = (float*)malloc(sizeof(float) * (critical_range_length));
     float* densities_resid_fit_on_critical_range = (float*)malloc(sizeof(float) * (critical_range_length));
     float* weights_for_downup = (float*)malloc(sizeof(float) * (critical_range_length));
-    for (int i = 0; i < critical_range_length; i++) {
+    for (bigint i = 0; i < critical_range_length; i++) {
         densities_resid_on_critical_range[i] = densities_resid[critical_range_min + i];
         weights_for_downup[i] = spacings[critical_range_min + i];
     }
     jisotonic5_downup(critical_range_length, densities_resid_fit_on_critical_range, densities_resid_on_critical_range, weights_for_downup);
 
-    int cutpoint_index = ns_isocut5::find_min_index(critical_range_length, densities_resid_fit_on_critical_range);
+    bigint cutpoint_index = ns_isocut5::find_min_index(critical_range_length, densities_resid_fit_on_critical_range);
     *cutpoint_out = (X_sub[critical_range_min + cutpoint_index] + X_sub[critical_range_min + cutpoint_index + 1]) / 2;
 
     free(samples_sorted);
@@ -107,7 +107,7 @@ void isocut5(double* dipscore_out, double* cutpoint_out, int N, float* samples, 
     free(densities_resid_fit_on_critical_range);
 }
 
-void isocut5_old(double* dipscore_out, double* cutpoint_out, int N, float* samples, isocut5_opts opts)
+void isocut5_old(double* dipscore_out, double* cutpoint_out, bigint N, float* samples, isocut5_opts opts)
 {
     float* samples_sorted = (float*)malloc(sizeof(float) * N);
 
@@ -118,33 +118,33 @@ void isocut5_old(double* dipscore_out, double* cutpoint_out, int N, float* sampl
         jisotonic5_sort(N, samples_sorted, samples);
 
     float num_bins_factor = 1;
-    int num_bins = ceil(sqrt(N * 1.0 / 2) * num_bins_factor);
+    bigint num_bins = ceil(sqrt(N * 1.0 / 2) * num_bins_factor);
 
-    int num_bins_1 = ceil(num_bins / 2);
-    int num_bins_2 = num_bins - num_bins_1;
-    int num_intervals = num_bins_1 + num_bins_2;
+    bigint num_bins_1 = ceil(num_bins / 2);
+    bigint num_bins_2 = num_bins - num_bins_1;
+    bigint num_intervals = num_bins_1 + num_bins_2;
     float* intervals = (float*)malloc(sizeof(float) * num_intervals);
-    for (int i = 0; i < num_bins_1; i++)
+    for (bigint i = 0; i < num_bins_1; i++)
         intervals[i] = i + 1;
-    for (int i = 0; i < num_bins_2; i++)
+    for (bigint i = 0; i < num_bins_2; i++)
         intervals[num_intervals - 1 - i] = i + 1;
     float alpha = (N - 1) / ns_isocut5::sum(num_intervals, intervals);
-    for (int i = 0; i < num_intervals; i++)
+    for (bigint i = 0; i < num_intervals; i++)
         intervals[i] *= alpha;
-    int N_sub = num_intervals + 1;
+    bigint N_sub = num_intervals + 1;
     float* inds = (float*)malloc(sizeof(float) * N_sub);
     inds[0] = 0;
-    for (int i = 0; i < num_intervals; i++)
+    for (bigint i = 0; i < num_intervals; i++)
         inds[i + 1] = inds[i] + intervals[i];
     float* X_sub = (float*)malloc(sizeof(float) * N_sub);
-    for (int i = 0; i < N_sub; i++)
-        X_sub[i] = samples_sorted[(int)inds[i]];
+    for (bigint i = 0; i < N_sub; i++)
+        X_sub[i] = samples_sorted[(bigint)inds[i]];
     float* densities = (float*)malloc(sizeof(float) * (N_sub - 1));
     float* spacings = (float*)malloc(sizeof(float) * (N_sub - 1));
     float* multiplicities = (float*)malloc(sizeof(float) * (N_sub - 1));
-    for (int i = 0; i < N_sub - 1; i++) {
+    for (bigint i = 0; i < N_sub - 1; i++) {
         spacings[i] = X_sub[i + 1] - X_sub[i];
-        multiplicities[i] = ((int)inds[i + 1]) - ((int)inds[i]);
+        multiplicities[i] = ((bigint)inds[i + 1]) - ((bigint)inds[i]);
         densities[i] = multiplicities[i] / spacings[i];
     }
 
@@ -154,17 +154,17 @@ void isocut5_old(double* dipscore_out, double* cutpoint_out, int N, float* sampl
     float* weights_for_downup = (float*)malloc(sizeof(float) * (N_sub - 1));
 
     jisotonic5_updown(N_sub - 1, densities_unimodal_fit, densities, multiplicities);
-    for (int i = 0; i < N_sub - 1; i++)
+    for (bigint i = 0; i < N_sub - 1; i++)
         densities_resid[i] = densities[i] - densities_unimodal_fit[i];
-    for (int i = 0; i < N_sub - 1; i++)
+    for (bigint i = 0; i < N_sub - 1; i++)
         weights_for_downup[i] = spacings[i];
     jisotonic5_downup(N_sub - 1, densities_resid_fit, densities_resid, weights_for_downup);
 
-    int cutpoint_index = ns_isocut5::find_min_index(N_sub - 1, densities_resid_fit);
+    bigint cutpoint_index = ns_isocut5::find_min_index(N_sub - 1, densities_resid_fit);
     *cutpoint_out = (X_sub[cutpoint_index] + X_sub[cutpoint_index + 1]) / 2;
 
     float* densities_unimodal_fit_times_spacings = (float*)malloc(sizeof(float) * (N_sub - 1));
-    for (int i = 0; i < N_sub - 1; i++)
+    for (bigint i = 0; i < N_sub - 1; i++)
         densities_unimodal_fit_times_spacings[i] = densities_unimodal_fit[i] * spacings[i];
     *dipscore_out = ns_isocut5::compute_ks4(N_sub - 1, multiplicities, densities_unimodal_fit_times_spacings);
 
@@ -184,41 +184,41 @@ void isocut5_old(double* dipscore_out, double* cutpoint_out, int N, float* sampl
 
 namespace ns_isocut5 {
 
-void copy_samples(int N, float* out, float* in)
+void copy_samples(bigint N, float* out, float* in)
 {
-    for (int i = 0; i < N; i++)
+    for (bigint i = 0; i < N; i++)
         out[i] = in[i];
 }
 
-double sum(int N, float* X)
+double sum(bigint N, float* X)
 {
     double ret = 0;
-    for (int i = 0; i < N; i++)
+    for (bigint i = 0; i < N; i++)
         ret += X[i];
     return ret;
 }
 
-int find_min_index(int N, float* X)
+bigint find_min_index(bigint N, float* X)
 {
-    int ret = 0;
-    for (int i = 0; i < N; i++) {
+    bigint ret = 0;
+    for (bigint i = 0; i < N; i++) {
         if (X[i] < X[ret])
             ret = i;
     }
     return ret;
 }
 
-int find_max_index(int N, float* X)
+bigint find_max_index(bigint N, float* X)
 {
-    int ret = 0;
-    for (int i = 0; i < N; i++) {
+    bigint ret = 0;
+    for (bigint i = 0; i < N; i++) {
         if (X[i] > X[ret])
             ret = i;
     }
     return ret;
 }
 
-double compute_ks4(int N, float* counts1, float* counts2)
+double compute_ks4(bigint N, float* counts1, float* counts2)
 {
     double sum_counts1 = sum(N, counts1);
     double sum_counts2 = sum(N, counts2);
@@ -227,7 +227,7 @@ double compute_ks4(int N, float* counts1, float* counts2)
     double cumsum_counts2 = 0;
 
     double max_diff = 0;
-    for (int i = 0; i < N; i++) {
+    for (bigint i = 0; i < N; i++) {
         cumsum_counts1 += counts1[i];
         cumsum_counts2 += counts2[i];
         double diff = fabs(cumsum_counts1 / sum_counts1 - cumsum_counts2 / sum_counts2);
@@ -238,7 +238,7 @@ double compute_ks4(int N, float* counts1, float* counts2)
     return max_diff * sqrt((sum_counts1 + sum_counts2) / 2);
 }
 
-double compute_ks5(int* critical_range_min, int* critical_range_max, int N, float* counts1, float* counts2, int peak_index)
+double compute_ks5(bigint* critical_range_min, bigint* critical_range_max, bigint N, float* counts1, float* counts2, bigint peak_index)
 {
     *critical_range_min = 0;
     *critical_range_max = N - 1; //should get over-written!
@@ -248,11 +248,11 @@ double compute_ks5(int* critical_range_min, int* critical_range_max, int N, floa
     {
         float* counts1_left = (float*)malloc(sizeof(float) * (peak_index + 1));
         float* counts2_left = (float*)malloc(sizeof(float) * (peak_index + 1));
-        for (int i = 0; i <= peak_index; i++) {
+        for (bigint i = 0; i <= peak_index; i++) {
             counts1_left[i] = counts1[i];
             counts2_left[i] = counts2[i];
         }
-        int len = peak_index + 1;
+        bigint len = peak_index + 1;
         while ((len >= 4) || (len == peak_index + 1)) {
             double ks0 = compute_ks4(len, counts1_left, counts2_left);
             if (ks0 > ks_best) {
@@ -268,11 +268,11 @@ double compute_ks5(int* critical_range_min, int* critical_range_max, int N, floa
     {
         float* counts1_right = (float*)malloc(sizeof(float) * (N - peak_index));
         float* counts2_right = (float*)malloc(sizeof(float) * (N - peak_index));
-        for (int i = 0; i < N - peak_index; i++) {
+        for (bigint i = 0; i < N - peak_index; i++) {
             counts1_right[i] = counts1[N - 1 - i];
             counts2_right[i] = counts2[N - 1 - i];
         }
-        int len = N - peak_index;
+        bigint len = N - peak_index;
         while ((len >= 4) || (len == N - peak_index)) {
             double ks0 = compute_ks4(len, counts1_right, counts2_right);
             if (ks0 > ks_best) {
@@ -287,9 +287,9 @@ double compute_ks5(int* critical_range_min, int* critical_range_max, int N, floa
     return ks_best;
 }
 
-void debug_print_array(int N, float* X)
+void debug_print_array(bigint N, float* X)
 {
-    for (int i = 0; i < N; i++) {
+    for (bigint i = 0; i < N; i++) {
         if ((i > 0) && (i % 10 == 0))
             printf("\n");
         printf("%g ", X[i]);
