+++ mountainsort/src/isosplit5/isocut5.h
@@ -7,7 +7,7 @@ struct isocut5_opts {
     bool already_sorted = false;
 };
 
-void isocut5(double* dipscore_out, double* cutpoint_out, int N, float* samples, isocut5_opts opts);
+void isocut5(double* dipscore_out, double* cutpoint_out, bigint N, float* samples, isocut5_opts opts);
 
 /*
  * MCWRAP [ dipscore[1,1], cutpoint[1,1] ] = isocut5_mex(samples[1,N])
